requirements.txt
streamlit
requests
pyttsx3
pandas
google-genai
fpdf2
flask
